Refactor module imports for consistency and clarity across the codebase
