+++ braintrace/_etrace_algorithms.py
@@ -23,7 +23,7 @@ from typing import Dict, Any, Optional
 
 import brainstate
 
-from ._etrace_compiler_graph import ETraceGraph
+from braintrace._etrace_compiler import ETraceGraph
 from ._etrace_graph_executor import ETraceGraphExecutor
 from ._typing import Path
