+++ braintrace/_etrace_compiler/hidden_group.py
@@ -38,7 +38,7 @@
 #             and compiled states (brainstate.transform.StatefulFunction)
 #       - [x] add the support for the "HiddenGroupState" and "ETraceTreeState"
 #       - [x] add the support for the "ElemWiseParam"
-#       - [x] split into "_etrace_compiler.py", "_etrace_vjp_compiler_graph.py", and "_etrace_compiler_hidden_group.py",
+#       - [x] split into "_etrace_compiler.py", "_etrace_vjp_compiler_graph.py", and "hidden_group.py",
 #
 # ==============================================================================
 
@@ -53,11 +53,11 @@ import jax.core
 import numpy as np
 from brainstate import HiddenGroupState
 
-from ._compatible_imports import Var, Literal, JaxprEqn, Jaxpr
-from ._etrace_compiler_base import JaxprEvaluation, find_matched_vars
-from ._etrace_compiler_module_info import extract_module_info, ModuleInfo
-from ._misc import NotSupportedError
-from ._typing import (
+from braintrace._compatible_imports import Var, Literal, JaxprEqn, Jaxpr
+from .base import JaxprEvaluation, find_matched_vars
+from .module_info import extract_module_info, ModuleInfo
+from braintrace._misc import NotSupportedError
+from braintrace._typing import (
     PyTree,
     HiddenInVar,
     HiddenOutVar,
