+++ braintrace/_etrace_compiler/module_info_test.py
