+++ braintrace/_etrace_vjp/pp_prop.py
@@ -34,8 +34,8 @@ import jax
 import jax.numpy as jnp
 
 from braintrace._etrace_algorithms import EligibilityTrace
-from braintrace._etrace_compiler_hid_param_op import HiddenParamOpRelation
-from braintrace._etrace_compiler_hidden_group import HiddenGroup
+from braintrace._etrace_compiler import HiddenGroup
+from braintrace._etrace_compiler import HiddenParamOpRelation
 from braintrace._etrace_concepts import ElemWiseParam
 from braintrace._misc import (
     check_dict_keys,
